Refactor BufferRenderTarget to be more testable
