+++ src/renderers/gl/CMakeLists.txt
@@ -28,7 +28,7 @@ ADD_LIBRARY(
 
   renderer.cpp
   renderer_factory.cpp
-  buffer_render_target.cpp
+  basic_buffer_render_target.cpp
 )
 
 target_link_libraries(mirrenderergl
