+++ src/server/compositor/default_configuration.cpp
@@ -27,7 +27,7 @@
 #include "mir/main_loop.h"
 #include "mir/graphics/display.h"
 #include "mir/system_executor.h"
-#include "mir/renderer/gl/buffer_render_target.h"
+#include "mir/renderer/gl/basic_buffer_render_target.h"
 #include "mir/renderer/gl/context.h"
 #include "mir/renderer/renderer.h"
 
@@ -100,7 +100,7 @@ auto mir::DefaultServerConfiguration::the_screen_shooter() -> std::shared_ptr<co
     return screen_shooter(
         [this]()
         {
-            auto render_target = std::make_unique<mrg::BufferRenderTarget>(the_display()->create_gl_context());
+            auto render_target = std::make_unique<mrg::BasicBufferRenderTarget>(the_display()->create_gl_context());
             auto renderer = the_renderer_factory()->create_renderer_for(*render_target);
             return std::make_shared<compositor::BasicScreenShooter>(
                 the_scene(),
